layers: Create shader_stage_state.cpp
