+++ layers/state_tracker/state_tracker.cpp
@@ -29,8 +29,8 @@
 #include "generated/chassis.h"
 #include "generated/state_tracker_helper.h"
 #include "state_tracker/state_tracker.h"
-#include "utils/shader_utils.h"
 #include "sync/sync_utils.h"
+#include "state_tracker/shader_stage_state.h"
 #include "state_tracker/image_state.h"
 #include "state_tracker/buffer_state.h"
 #include "state_tracker/device_state.h"
