+++ layers/utils/shader_utils.h
@@ -24,21 +24,10 @@
 #include "utils/vk_layer_utils.h"
 
 #include <spirv-tools/libspirv.hpp>
-#include <vulkan/utility/vk_safe_struct.hpp>
 
 struct DeviceFeatures;
 struct DeviceExtensions;
 class APIVersion;
-
-namespace spirv {
-struct ResourceInterfaceVariable;
-}  // namespace spirv
-
-struct DescriptorRequirement {
-    uint64_t revalidate_hash;
-    const spirv::ResourceInterfaceVariable *variable;
-    DescriptorRequirement() : revalidate_hash(0), variable(nullptr) {}
-};
 
 enum class ShaderObjectStage : uint32_t {
     VERTEX = 0u,
@@ -78,59 +67,6 @@ inline ShaderObjectStage VkShaderStageToShaderObjectStage(VkShaderStageFlagBits 
     }
     return ShaderObjectStage::LAST;
 }
-
-inline bool operator==(const DescriptorRequirement &a, const DescriptorRequirement &b) noexcept {
-    return a.revalidate_hash == b.revalidate_hash;
-}
-
-inline bool operator<(const DescriptorRequirement &a, const DescriptorRequirement &b) noexcept {
-    return a.revalidate_hash < b.revalidate_hash;
-}
-
-// < binding index (of descriptor set) : meta data >
-typedef std::unordered_multimap<uint32_t, DescriptorRequirement> BindingVariableMap;
-
-// Capture which slots (set#->bindings) are actually used by the shaders of this pipeline
-using ActiveSlotMap = vvl::unordered_map<uint32_t, BindingVariableMap>;
-
-namespace vku {
-namespace safe {
-struct PipelineShaderStageCreateInfo;
-struct ShaderCreateInfoEXT;
-struct SpecializationInfo;
-}  // namespace safe
-}  // namespace vku
-
-namespace vvl {
-struct ShaderModule;
-}  // namespace vvl
-
-namespace spirv {
-struct Module;
-struct EntryPoint;
-class Instruction;
-}  // namespace spirv
-
-// This is a wrapper around the Shader as it may come from a Pipeline or Shader Object.
-struct ShaderStageState {
-    // We use this over a spirv::Module because there are times we need to create empty objects
-    std::shared_ptr<const vvl::ShaderModule> module_state;
-    std::shared_ptr<const spirv::Module> spirv_state;
-    const vku::safe_VkPipelineShaderStageCreateInfo *pipeline_create_info;
-    const vku::safe_VkShaderCreateInfoEXT *shader_object_create_info;
-    // If null, means it is an empty object, no SPIR-V backing it
-    std::shared_ptr<const spirv::EntryPoint> entrypoint;
-
-    ShaderStageState(const vku::safe_VkPipelineShaderStageCreateInfo *pipeline_create_info,
-                     const vku::safe_VkShaderCreateInfoEXT *shader_object_create_info,
-                     std::shared_ptr<const vvl::ShaderModule> module_state, std::shared_ptr<const spirv::Module> spirv_state);
-
-    const char *GetPName() const;
-    VkShaderStageFlagBits GetStage() const;
-    vku::safe_VkSpecializationInfo *GetSpecializationInfo() const;
-    const void *GetPNext() const;
-    bool GetInt32ConstantValue(const spirv::Instruction &insn, uint32_t *value) const;
-};
 
 class ValidationCache {
   public:
@@ -178,8 +114,3 @@ spv_target_env PickSpirvEnv(const APIVersion &api_version, bool spirv_1_4);
 void AdjustValidatorOptions(const DeviceExtensions &device_extensions, const DeviceFeatures &enabled_features,
                             spvtools::ValidatorOptions &out_options, uint32_t *out_hash);
 
-void GetActiveSlots(ActiveSlotMap &active_slots, const std::shared_ptr<const spirv::EntryPoint> &entrypoint);
-ActiveSlotMap GetActiveSlots(const std::vector<ShaderStageState> &stage_states);
-ActiveSlotMap GetActiveSlots(const std::shared_ptr<const spirv::EntryPoint> &entrypoint);
-
-uint32_t GetMaxActiveSlot(const ActiveSlotMap &active_slots);
